backend/requirements.txt
# ==========================================
# PRODUCCIÓN - Dependencias para Lambda
# ==========================================

# Conversión de formatos
markdown==3.5.1
python-docx==1.1.0
htmldocx==0.0.6

# AWS SDK
boto3==1.34.10
botocore==1.34.10

# Utilidades
python-dateutil==2.8.2